+++ contracts/cosmwasm-vm/cw-ibc-core/src/ics04_channel/handler.rs
@@ -103,7 +103,7 @@ impl<'a> ValidateChannel for CwIbcCoreContext<'a> {
             EXECUTE_ON_CHANNEL_OPEN_INIT,
             &sub_message.channel().endpoint,
         )?;
-        let data = cw_common::xcall_connection_msg::ExecuteMsg::IbcChannelOpen { msg: sub_message };
+        let data = cw_common::ibc_dapp_msg::ExecuteMsg::IbcChannelOpen { msg: sub_message };
         let data = to_binary(&data).map_err(ContractError::Std)?;
         let on_chan_open_init = create_channel_submesssage(
             contract_address,
@@ -240,7 +240,7 @@ impl<'a> ValidateChannel for CwIbcCoreContext<'a> {
             &sub_message.channel().endpoint,
         )?;
 
-        let data = cw_common::xcall_connection_msg::ExecuteMsg::IbcChannelOpen { msg: sub_message };
+        let data = cw_common::ibc_dapp_msg::ExecuteMsg::IbcChannelOpen { msg: sub_message };
 
         let data = to_binary(&data).map_err(ContractError::Std)?;
         cw_println!(deps, "after converting data to binary ");
@@ -356,7 +356,7 @@ impl<'a> ValidateChannel for CwIbcCoreContext<'a> {
             &sub_message.channel().endpoint,
         )?;
         let data =
-            cw_common::xcall_connection_msg::ExecuteMsg::IbcChannelConnect { msg: sub_message };
+            cw_common::ibc_dapp_msg::ExecuteMsg::IbcChannelConnect { msg: sub_message };
         let data = to_binary(&data).unwrap();
         let on_chan_open_try = create_channel_submesssage(
             module_contract_address,
@@ -470,7 +470,7 @@ impl<'a> ValidateChannel for CwIbcCoreContext<'a> {
             &sub_message.channel().endpoint,
         )?;
         let data =
-            cw_common::xcall_connection_msg::ExecuteMsg::IbcChannelConnect { msg: sub_message };
+            cw_common::ibc_dapp_msg::ExecuteMsg::IbcChannelConnect { msg: sub_message };
         let data = to_binary(&data).unwrap();
         let on_chan_open_try = create_channel_submesssage(
             contract_address,
@@ -529,7 +529,7 @@ impl<'a> ValidateChannel for CwIbcCoreContext<'a> {
             &sub_message.channel().endpoint,
         )?;
         let data =
-            cw_common::xcall_connection_msg::ExecuteMsg::IbcChannelClose { msg: sub_message };
+            cw_common::ibc_dapp_msg::ExecuteMsg::IbcChannelClose { msg: sub_message };
         let data = to_binary(&data).unwrap();
         let on_chan_close_init = create_channel_submesssage(
             contract_address,
@@ -642,7 +642,7 @@ impl<'a> ValidateChannel for CwIbcCoreContext<'a> {
             &sub_message.channel().endpoint,
         )?;
         let data =
-            cw_common::xcall_connection_msg::ExecuteMsg::IbcChannelClose { msg: sub_message };
+            cw_common::ibc_dapp_msg::ExecuteMsg::IbcChannelClose { msg: sub_message };
         let data = to_binary(&data).map_err(Into::<ContractError>::into)?;
         let on_chan_close_confirm = create_channel_submesssage(
             contract_address,
