+++ contracts/cosmwasm-vm/cw-ibc-core/src/ics04_channel/packet/receive_packet.rs
@@ -182,7 +182,7 @@ impl<'a> CwIbcCoreContext<'a> {
             VALIDATE_ON_PACKET_RECEIVE_ON_MODULE,
             &ibc_packet,
         )?;
-        let cosm_msg = cw_common::xcall_connection_msg::ExecuteMsg::IbcPacketReceive {
+        let cosm_msg = cw_common::ibc_dapp_msg::ExecuteMsg::IbcPacketReceive {
             msg: cosmwasm_std::IbcPacketReceiveMsg::new(ibc_packet, address),
         };
         let create_client_message: CosmosMsg = CosmosMsg::Wasm(cosmwasm_std::WasmMsg::Execute {
