+++ contracts/cosmwasm-vm/cw-ibc-core/tests/channel/test_channel_closeinit.rs
@@ -86,7 +86,7 @@ fn test_validate_close_init_channel() {
         )
         .unwrap();
 
-    let data = cw_common::xcall_connection_msg::ExecuteMsg::IbcChannelClose { msg: expected };
+    let data = cw_common::ibc_dapp_msg::ExecuteMsg::IbcChannelClose { msg: expected };
     let data = to_binary(&data).unwrap();
     let on_chan_open_init = create_channel_submesssage(
         "contractaddress".to_string(),
