+++ contracts/cosmwasm-vm/cw-integration/tests/test_host_lightclient.rs
@@ -310,7 +310,7 @@ pub fn call_set_xcall_host(ctx: &mut TestContext) -> Result<AppResponse, AppErro
     ctx.app.execute_contract(
         ctx.sender.clone(),
         ctx.get_xcall_ibc_connection(),
-        &cw_common::xcall_connection_msg::ExecuteMsg::SetXCallHost {
+        &cw_common::ibc_dapp_msg::ExecuteMsg::SetXCallHost {
             address: ctx.get_xcall_app().to_string(),
         },
         &[],
@@ -326,7 +326,7 @@ pub fn call_configure_connection(
     ctx.app.execute_contract(
         ctx.sender.clone(),
         ctx.get_xcall_ibc_connection(),
-        &cw_common::xcall_connection_msg::ExecuteMsg::ConfigureConnection {
+        &cw_common::ibc_dapp_msg::ExecuteMsg::ConfigureConnection {
             connection_id,
             counterparty_port_id: PORT.to_string(),
             counterparty_nid: NetId::from(nid),
