+++ contracts/cosmwasm-vm/cw-xcall-ibc-connection/tests/test_admin.rs
@@ -2,7 +2,7 @@ mod account;
 mod setup;
 use account::*;
 use cosmwasm_std::testing::mock_env;
-use cw_common::xcall_connection_msg::QueryMsg;
+use cw_common::ibc_dapp_msg::QueryMsg;
 use cw_xcall_ibc_connection::state::CwIbcConnection;
 use setup::*;
 
@@ -258,7 +258,7 @@ fn add_invalid_char_as_admin() {
             mock_deps.as_mut(),
             mock_env,
             mock_info,
-            cw_common::xcall_connection_msg::ExecuteMsg::SetAdmin {
+            cw_common::ibc_dapp_msg::ExecuteMsg::SetAdmin {
                 address: "*************".into(),
             },
         )
@@ -286,7 +286,7 @@ fn validate_address_add_admin_size_lessthan_3() {
             mock_deps.as_mut(),
             mock_env,
             mock_info,
-            cw_common::xcall_connection_msg::ExecuteMsg::SetAdmin {
+            cw_common::ibc_dapp_msg::ExecuteMsg::SetAdmin {
                 address: "sm".into(),
             },
         )
@@ -314,7 +314,7 @@ fn validate_address_add_admin_size_more_than_45() {
             mock_deps.as_mut(),
             mock_env,
             mock_info,
-            cw_common::xcall_connection_msg::ExecuteMsg::SetAdmin {
+            cw_common::ibc_dapp_msg::ExecuteMsg::SetAdmin {
                 address: "eddiuo6lbp05golmz3rb5n7hbi4c5hhyh0rb1w6cslyjt5mhwd0chn3x254lyorpx4dzvrvsc9h2em44be2rj193dwe".into(),
             },
         )
